Add authorization

+++ app/server/src/main/java/com/bsuir/tracker/Service/EmployeeServiceImpl.java
@@ -1,6 +1,7 @@
 package com.bsuir.tracker.Service;
 
 import com.bsuir.tracker.DAO.EmployeeDAO;
+import com.bsuir.tracker.DAO.RoleDAO;
 import com.bsuir.tracker.entity.EmployeeEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,6 +18,11 @@ import java.util.List;
 public class EmployeeServiceImpl implements EmployeeService {
     @Autowired
     private EmployeeDAO employeeDAO;
+    @Autowired
+    private RoleDAO roleDAO;
+
+    private  String ADMINISTRATOR_ROLE_IDENTIFIER = "26";
+    private  String DIRECTOR_ROLE_IDENTIFIER = "31";
 
     @Transactional
     public void addEmployee(EmployeeEntity employee) {
@@ -26,6 +32,8 @@ public class EmployeeServiceImpl implements EmployeeService {
     public EmployeeEntity getEmployee(int idEmployee) {
         return employeeDAO.getEmployee(idEmployee);
     }
+
+    public EmployeeEntity getEmployeeByMail(String mail) {return employeeDAO.getEmployeeByMail(mail); }
 
     @Transactional
     public List<EmployeeEntity> getAllEmployees() {
@@ -38,7 +46,26 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Transactional
     public void deleteEmployee(int idEmployee) {
-        employeeDAO.deleteEmployee(idEmployee);
+        if (countAdministrators() > 1)
+        {
+            employeeDAO.deleteEmployee(idEmployee);
+        }
+    }
+
+    public int countAdministrators()
+    {
+        int counter = 0;
+        List<EmployeeEntity> employeeEntityList = employeeDAO.getAllEmployees();
+        if (employeeEntityList != null)
+            for(EmployeeEntity employeeEntity : employeeEntityList)
+            {
+                String code = roleDAO.getRole(employeeEntity.getRoleIdrole()).getCode();
+                if (code.equals(ADMINISTRATOR_ROLE_IDENTIFIER))
+                {
+                    counter++;
+                }
+            }
+        return counter;
     }
 
     public void setEmployeeDAO(EmployeeDAO employeeDAO){
